+++ src/middleware/permission.js
@@ -7,9 +7,6 @@ import _ from 'lodash'
 export default async (ctx, next) => {
   let req = ctx.request
   let path = req.url
-  const ip = ctx.ip
-  //TODO:对客户端ip进行过滤等操作
-  // ctx.fpm.get('')
   //如果调用的是api路由，则进行分析和拦截
   if(path === '/api'){
     //验证安全性