+++ test/function.test.js
@@ -1,6 +1,6 @@
 var should = require("chai").should;
 var YF = require("yf-fpm-client-nodejs").default;
-YF.init({ appkey:'123123', masterKey:'123', endpoint: 'http://localhost:9999/api' });
+YF.init({ mode: 'DEV', appkey:'123123', masterKey:'123123', endpoint: 'http://localhost:9999/api' });
 
 describe('Function', function(){
   it('test', function(done){